Moved Navbar title to its own component
