+++ frontend/src/components/Navbar/Navbar.js
@@ -1,8 +1,8 @@
 import React from "react";
-import { Link } from "react-router-dom";
-import { AppBar, Toolbar, Container, Typography } from "@material-ui/core";
+import { AppBar, Toolbar, Container } from "@material-ui/core";
 
 import NavbarAuthButton from "./NavbarAuthButton";
+import NavbarTitle from "./NavbarTitle";
 
 const Navbar = () => {
   return (
@@ -10,11 +10,7 @@ const Navbar = () => {
       <AppBar>
         <Toolbar>
           <Container style={{ display: "flex" }}>
-            <Typography style={{ flexGrow: 1 }} variant="h6">
-              <Link style={{ textDecoration: "none", color: "white" }} to="/">
-                Meebook Homework Checker
-              </Link>
-            </Typography>
+            <NavbarTitle />
             <NavbarAuthButton />
           </Container>
         </Toolbar>
